// graphql/codegen/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Int or string */
  ActionCount: any;
  /** AreaOfEffectType ("SPHERE", "CUBE", "CYLINDER", "LINE", "CONE") or list of AreaOfEffectTypes */
  AreaOfEffectTypeFilter: any;
  /** Float, list of floats, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  FloatFilter: any;
  /** Int, list of ints, or object with gte (>=), gt (>), lte (<=), and lt (<) properties for range of values */
  IntFilter: any;
  /** LanguageScript ("COMMON", "ELVISH", "DWARVISH", "INFERNAL", "DRACONIC", "CELESTIAL") or list of LanguageScripts */
  LanguageScriptFilter: any;
  /** MonsterSubtype ("ANY_RACE", "HUMAN", "DWARF", "ELF", "GOBLINOID", "MERFOLK", "SHAPECHANGER", "DEMON", "DEVIL", "ORC", "SAHUAGIN", "TITAN", "KOBOLD", "GNOLL", "GRIMLOCK", "LIZARDFOLK", "GNOME") or list of MonsterSubtypes */
  MonsterSubtypeFilter: any;
  /** MonsterTypes ("BEAST", "MONSTROSITY", "DRAGON", "HUMANOID", "UNDEAD", "FIEND", "CELESTIAL", "CONSTRUCT", "GIANT", "ELEMENTAL", "FEY", "ABERRATION", "OOZE", "SWARM", "PLANT") or list of MonsterTypes */
  MonsterTypeFilter: any;
  /** ProficiencyType ("WEAPONS", "ARTISANS_TOOLS", "SKILLS", "ARMOR", "MUSICAL_INSTRUMENTS", "SAVING_THROWS", "OTHER", "GAMING_SETS", "VEHICLES") or list of ProficiencyTypes */
  ProficiencyTypeFilter: any;
  /** Size ("TINY", "SMALL", "MEDIUM", "LARGE", "HUGE", "GARGANTUAN") or list of sizes */
  SizeFilter: any;
  /** SpellAttackType ("MELEE", "RANGED") or list of SpellAttackTypes */
  SpellAttackTypeFilter: any;
  /** String or list of strings */
  StringFilter: any;
};
// ...
export type AbilityScore = {
  __typename?: 'AbilityScore';
  desc: Array<Scalars['String']>;
  full_name: Scalars['String'];
  index: Scalars['String'];
  name: Scalars['String'];
  skills: Array<Skill>;
};
// ...
export type AreaOfEffect = {
  __typename?: 'AreaOfEffect';
  size: Scalars['Int'];
  type: AreaOfEffectType;
};

export type AreaOfEffectFilter = {
  size?: InputMaybe<Scalars['IntFilter']>;
  type?: InputMaybe<Scalars['AreaOfEffectTypeFilter']>;
};

export enum AreaOfEffectType {
  Cone = 'CONE',
  Cube = 'CUBE',
  Cylinder = 'CYLINDER',
  Line = 'LINE',
  Sphere = 'SPHERE'
}
// ...
export type Class = {
  __typename?: 'Class';
  class_levels: Array<Level>;
  hit_die: Scalars['Int'];
  index: Scalars['String'];
  multi_classing: Multiclassing;
  name: Scalars['String'];
  proficiencies: Array<Proficiency>;
  proficiency_choices: Array<ProficiencyChoice>;
  saving_throws: Array<AbilityScore>;
  spellcasting?: Maybe<ClassSpellcasting>;
  spells?: Maybe<Array<Spell>>;
  starting_equipment: Array<Quantity>;
  starting_equipment_options: Array<StartingEquipmentChoice>;
  subclasses: Array<Subclass>;
};
// ...
export type ClassSpellcasting = {
  __typename?: 'ClassSpellcasting';
  info: Array<SpellcastingInfo>;
  level: Scalars['Int'];
  spellcasting_ability: AbilityScore;
};
// ...
export type DamageAtLevel = {
  __typename?: 'DamageAtLevel';
  damage: Scalars['String'];
  level: Scalars['Int'];
};

export type DamageOption = {
  __typename?: 'DamageOption';
  damage_dice: Scalars['String'];
  damage_type: DamageType;
  notes?: Maybe<Scalars['String']>;
  option_type: Scalars['String'];
};

export type DamageOptionSet = {
  __typename?: 'DamageOptionSet';
  option_set_type: Scalars['String'];
  options: Array<DamageOption>;
};

export type DamageType = {
  __typename?: 'DamageType';
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
};

export enum DcSuccess {
  Half = 'HALF',
  None = 'NONE',
  Other = 'OTHER'
}
// ...
export type HealingAtLevel = {
  __typename?: 'HealingAtLevel';
  healing: Scalars['String'];
  level: Scalars['Int'];
};
// ...
export type Level = {
  __typename?: 'Level';
  ability_score_bonuses?: Maybe<Scalars['Int']>;
  class: Class;
  class_specific?: Maybe<ClassSpecific>;
  features: Array<Feature>;
  index: Scalars['String'];
  level: Scalars['Int'];
  prof_bonus?: Maybe<Scalars['Int']>;
  spellcasting?: Maybe<LevelSpellcasting>;
  subclass?: Maybe<Subclass>;
  subclass_specific?: Maybe<SubclassSpecific>;
};
// ...
export type MagicSchool = {
  __typename?: 'MagicSchool';
  desc: Scalars['String'];
  index: Scalars['String'];
  name: Scalars['String'];
  spells: Array<Spell>;
};
// ...
export type Multiclassing = {
  __typename?: 'Multiclassing';
  prerequisite_options?: Maybe<PrerequisiteChoice>;
  prerequisites?: Maybe<Array<AbilityScorePrerequisite>>;
  proficiencies: Array<Proficiency>;
  proficiency_choices?: Maybe<Array<ProficiencyChoice>>;
};
// ...
export type Proficiency = {
  __typename?: 'Proficiency';
  classes: Array<Class>;
  index: Scalars['String'];
  name: Scalars['String'];
  races: Array<ProficiencyRace>;
  reference: ProficiencyReference;
  type: ProficiencyType;
};


export type ProficiencyClassesArgs = {
  name?: InputMaybe<Scalars['String']>;
};


export type ProficiencyRacesArgs = {
  name?: InputMaybe<Scalars['String']>;
};

export type ProficiencyChoice = {
  __typename?: 'ProficiencyChoice';
  choose: Scalars['Int'];
  desc?: Maybe<Scalars['String']>;
  from: ProficiencyOptionSet;
  type: Scalars['String'];
};
// ...
export type Quantity = {
  __typename?: 'Quantity';
  equipment: IEquipment;
  quantity: Scalars['Int'];
};
// ...
export type Skill = {
  __typename?: 'Skill';
  ability_score: AbilityScore;
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type Spell = {
  __typename?: 'Spell';
  area_of_effect?: Maybe<AreaOfEffect>;
  attack_type?: Maybe<SpellAttackType>;
  casting_time: Scalars['String'];
  classes: Array<Class>;
  components?: Maybe<Array<Maybe<SpellComponent>>>;
  concentration: Scalars['Boolean'];
  damage?: Maybe<SpellDamage>;
  dc?: Maybe<SpellDc>;
  desc: Array<Scalars['String']>;
  duration: Scalars['String'];
  heal_at_slot_level?: Maybe<Array<HealingAtLevel>>;
  higher_level?: Maybe<Array<Scalars['String']>>;
  index: Scalars['String'];
  level: Scalars['Int'];
  material?: Maybe<Scalars['String']>;
  name: Scalars['String'];
  range: Scalars['String'];
  ritual: Scalars['Boolean'];
  school: MagicSchool;
  subclasses: Array<Subclass>;
};


export type SpellClassesArgs = {
  name?: InputMaybe<Scalars['String']>;
};


export type SpellSubclassesArgs = {
  name?: InputMaybe<Scalars['String']>;
};

export enum SpellAttackType {
  Melee = 'MELEE',
  Ranged = 'RANGED'
}

export type SpellChoice = {
  __typename?: 'SpellChoice';
  choose: Scalars['Int'];
  from: SpellOptionSet;
  type: Scalars['String'];
};

export enum SpellComponent {
  M = 'M',
  S = 'S',
  V = 'V'
}

export type SpellDamage = {
  __typename?: 'SpellDamage';
  damage_at_character_level?: Maybe<Array<DamageAtLevel>>;
  damage_at_slot_level?: Maybe<Array<DamageAtLevel>>;
  damage_type?: Maybe<DamageType>;
};

export type SpellDc = {
  __typename?: 'SpellDc';
  desc?: Maybe<Scalars['String']>;
  success: DcSuccess;
  type: AbilityScore;
};
// ...
export type SpellWithPrerequisite = {
  __typename?: 'SpellWithPrerequisite';
  prerequisites: Array<Maybe<SpellPrerequisite>>;
  spell: Spell;
};

export type SpellcastingInfo = {
  __typename?: 'SpellcastingInfo';
  desc: Array<Scalars['String']>;
  name: Scalars['String'];
};

export type StartingEquipmentChoice = {
  __typename?: 'StartingEquipmentChoice';
  choose: Scalars['Int'];
  desc: Scalars['String'];
  from: StartingEquipmentOptionSet;
  type: Scalars['String'];
};
// ...
export type Subclass = {
  __typename?: 'Subclass';
  class: Class;
  desc: Array<Scalars['String']>;
  index: Scalars['String'];
  name: Scalars['String'];
  spells?: Maybe<Array<SpellWithPrerequisite>>;
  subclass_flavor: Scalars['String'];
  subclass_levels: Array<Maybe<Level>>;
};